+++ crates/module-system/sov-modules-api/src/runtime/capabilities/authentication.rs
@@ -5,21 +5,19 @@ use std::marker::PhantomData;
 use borsh::{BorshDeserialize, BorshSerialize};
 use serde::{Deserialize, Serialize};
 use sov_modules_macros::config_value_private;
-use sov_rollup_interface::crypto::CredentialId;
-use sov_rollup_interface::da::DaSpec;
 use sov_rollup_interface::TxHash;
 use sov_state::User;
 use thiserror::Error;
 
+use crate::capabilities::{AuthorizationData, UniquenessData};
 use crate::transaction::{
     AuthenticatedTransactionAndRawHash, Credentials, Transaction, TransactionVerificationError,
     VersionedTx,
 };
 use crate::{
-    capabilities, metered_credential, AuthenticatedTransactionData, Context, CryptoSpec,
-    DispatchCall, FullyBakedTx, GasMeter, GasMeteringError, MeteredBorshDeserialize,
+    capabilities, metered_credential, AuthenticatedTransactionData, CryptoSpec, DispatchCall,
+    FullyBakedTx, GasMeter, GasMeteringError, MeteredBorshDeserialize,
     MeteredBorshDeserializeError, MeteredHasher, ProvableStateReader, RawTx, Runtime, Spec,
-    StateAccessor,
 };
 
 /// The chain ID of the rollup.
@@ -172,42 +170,6 @@ where
     }
 }
 
-/// Authorizes transactions to be executed.
-pub trait TransactionAuthorizer<S: Spec> {
-    /// Resolves the [`Context`] for a transaction.
-    fn resolve_context(
-        &mut self,
-        auth_data: &AuthorizationData<S>,
-        sequencer: &<<S as Spec>::Da as DaSpec>::Address,
-        sequencer_rollup_address: S::Address,
-        state: &mut impl StateAccessor,
-    ) -> anyhow::Result<Context<S>>;
-
-    /// Resolves the context for an unregistered transaction.
-    fn resolve_unregistered_context(
-        &mut self,
-        auth_data: &AuthorizationData<S>,
-        sequencer: &<<S as Spec>::Da as DaSpec>::Address,
-        state: &mut impl StateAccessor,
-    ) -> anyhow::Result<Context<S>>;
-
-    /// Prevents duplicate transactions from running.
-    fn check_uniqueness(
-        &self,
-        auth_data: &AuthorizationData<S>,
-        context: &Context<S>,
-        state: &mut impl StateAccessor,
-    ) -> anyhow::Result<()>;
-
-    /// Marks a transaction as having been executed, preventing it from executing again.
-    fn mark_tx_attempted(
-        &mut self,
-        auth_data: &AuthorizationData<S>,
-        sequencer: &<<S as Spec>::Da as DaSpec>::Address,
-        state: &mut impl StateAccessor,
-    ) -> anyhow::Result<()>;
-}
-
 /// Output of the authentication.
 pub type AuthenticationOutput<S, Decodable> = (
     AuthenticatedTransactionAndRawHash<S>,
@@ -270,36 +232,6 @@ pub enum UnregisteredAuthenticationError {
     /// The transaction authentication failed because the authentication discriminant was invalid.
     #[error("Invalid authentication discriminant")]
     InvalidAuthenticationDiscriminant,
-}
-
-/// The different types of data that can be used to verify transaction uniqueness
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum UniquenessData {
-    /// Nonce-based uniqueness: an account's transactions must have a unique and consecutive nonces
-    Nonce(u64),
-    /// Generation-based uniqueness: the last `PAST_TRANSACTION_GENERATION` generations are cached.
-    /// Transactions older than this buffer are invalid, transactions falling within it or with a
-    /// higher generation are valid but must have a unique hash within their generation
-    Generation(u64),
-}
-
-/// Data required to authorize a sov-transaction.
-pub struct AuthorizationData<S: Spec> {
-    /// The nonce of the transaction.
-    pub uniqueness: UniquenessData,
-
-    /// The hash of the transaction.
-    pub tx_hash: TxHash,
-
-    /// Credential identifier used to retrieve relevant rollup address.
-    pub credential_id: CredentialId,
-
-    /// Holds the original credentials to authenticate the transaction and
-    /// provides information about which `Authenticator` was used to authenticate the transaction.
-    pub credentials: Credentials,
-
-    /// The default address.
-    pub default_address: S::Address,
 }
 
 fn verify_and_decode_tx<S: Spec, D: DispatchCall<Spec = S>>(
